ADMM - first try

+++ admm/admm_pytorch.py
@@ -48,7 +48,7 @@ if __name__ == '__main__':
             torch.manual_seed(42)
             X = torch.randn(n, dx).type(dtype)
             W = torch.randn(dy, dx).t().type(dtype)
-            Y = (torch.mm(X, W) + 10 * torch.randn(n, dy)).type(dtype)
+            Y = torch.mm(X, W).type(dtype) + 10 * torch.randn(n, dy).type(dtype)
 
             X_t_X, LU = torch.gesv(torch.eye(dx), torch.mm(X.t(), X))
             W_full_data = torch.mm(X_t_X, torch.mm(X.t(), Y))
